refactor: move functions into userControllerUtils file
